+++ packages/dodo-ui-bits/src/lib/index.ts
@@ -90,7 +90,14 @@ export {
 export {
   default as ModalManager,
   type ModalManagerProps,
+  type ModalManagerCustomCardBaseContext,
 } from './components/Feedback/Modals/ModalManager/ModalManager.svelte';
+
+export { type ModalManagerCustomCardModalContext } from './components/Feedback/Modals/ModalManager/ModalRenderer.svelte';
+
+export { type ModalManagerCustomCardInformContext } from './components/Feedback/Modals/ModalManager/InformRenderer.svelte';
+
+export { type ModalManagerCustomCardConfirmContext } from './components/Feedback/Modals/ModalManager/ConfirmRenderer.svelte';
 
 export {
   modals,